+++ frontend/weapp/pages/splash/index.js
@@ -1,0 +1,15 @@
+const { splashImage } = require('../../utils/constants');
+
+Page({
+  data: {
+    splashImage
+  },
+
+  onLoad() {
+    getApp().ensureLogin().catch(() => {});
+  },
+
+  start() {
+    wx.switchTab({ url: '/pages/home/index' });
+  }
+});